Update e2e test suite to not use new fields
