+++ test/e2e/suite/issuers/acme/certificate/webhook.go
@@ -34,7 +34,7 @@ import (
 	"github.com/jetstack/cert-manager/test/e2e/framework/addon/tiller"
 	"github.com/jetstack/cert-manager/test/e2e/framework/log"
 	"github.com/jetstack/cert-manager/test/e2e/util"
-	"github.com/jetstack/cert-manager/test/util/generate"
+	"github.com/jetstack/cert-manager/test/unit/gen"
 )
 
 var _ = framework.CertManagerDescribe("ACME webhook DNS provider", func() {
@@ -42,9 +42,6 @@ var _ = framework.CertManagerDescribe("ACME webhook DNS provider", func() {
 	//h := f.Helper()
 
 	Context("with the sample webhook solver deployed", func() {
-		// TODO: add additional DNS provider configs here
-		//cf := &dnsproviders.Cloudflare{}
-
 		var (
 			tiller = &tiller.Tiller{
 				Name:               "tiller-deploy-sample-webhook",
@@ -81,28 +78,31 @@ var _ = framework.CertManagerDescribe("ACME webhook DNS provider", func() {
 			dnsDomain = "example.com"
 
 			By("Creating an Issuer")
-			issuer := generate.Issuer(generate.IssuerConfig{
-				Name:               issuerName,
-				Namespace:          f.Namespace.Name,
-				ACMESkipTLSVerify:  true,
-				ACMEServer:         pebble.Details().Host,
-				ACMEEmail:          testingACMEEmail,
-				ACMEPrivateKeyName: testingACMEPrivateKey,
-				DNS01: &v1alpha1.ACMEIssuerDNS01Config{
-					Providers: []v1alpha1.ACMEIssuerDNS01Provider{
+			issuer := gen.Issuer(issuerName,
+				gen.SetIssuerACME(v1alpha1.ACMEIssuer{
+					SkipTLSVerify: true,
+					Server:        pebble.Details().Host,
+					Email:         testingACMEEmail,
+					PrivateKey: v1alpha1.SecretKeySelector{
+						LocalObjectReference: v1alpha1.LocalObjectReference{
+							Name: testingACMEPrivateKey,
+						},
+					},
+					Solvers: []v1alpha1.ACMEChallengeSolver{
 						{
-							Name: "default",
-							Webhook: &v1alpha1.ACMEIssuerDNS01ProviderWebhook{
-								GroupName:  webhook.Details().GroupName,
-								SolverName: webhook.Details().SolverName,
-								Config: &v1beta1.JSON{
-									Raw: []byte(`{}`),
+							DNS01: &v1alpha1.ACMEChallengeSolverDNS01{
+								Webhook: &v1alpha1.ACMEIssuerDNS01ProviderWebhook{
+									GroupName:  webhook.Details().GroupName,
+									SolverName: webhook.Details().SolverName,
+									Config: &v1beta1.JSON{
+										Raw: []byte(`{}`),
+									},
 								},
 							},
 						},
 					},
-				},
-			})
+				}))
+			issuer.Namespace = f.Namespace.Name
 			issuer, err := f.CertManagerClientSet.CertmanagerV1alpha1().Issuers(f.Namespace.Name).Create(issuer)
 			Expect(err).NotTo(HaveOccurred())
 			By("Waiting for Issuer to become Ready")
@@ -143,18 +143,13 @@ var _ = framework.CertManagerDescribe("ACME webhook DNS provider", func() {
 
 			certClient := f.CertManagerClientSet.CertmanagerV1alpha1().Certificates(f.Namespace.Name)
 
-			cert := generate.Certificate(generate.CertificateConfig{
-				Name:       certificateName,
-				Namespace:  f.Namespace.Name,
-				SecretName: certificateSecretName,
-				IssuerName: issuerName,
-				DNSNames:   []string{dnsDomain},
-				SolverConfig: &v1alpha1.SolverConfig{
-					DNS01: &v1alpha1.DNS01SolverConfig{
-						Provider: "default",
-					},
-				},
-			})
+			cert := gen.Certificate(certificateName,
+				gen.SetCertificateSecretName(certificateSecretName),
+				gen.SetCertificateIssuer(v1alpha1.ObjectReference{Name: issuerName}),
+				gen.SetCertificateDNSNames(dnsDomain),
+			)
+			cert.Namespace = f.Namespace.Name
+
 			cert, err := certClient.Create(cert)
 			Expect(err).NotTo(HaveOccurred())
