+++ test/e2e/suite/issuers/acme/dnsproviders/cloudflare.go
@@ -82,18 +82,6 @@ func (b *Cloudflare) Provision() error {
 	}
 
 	b.createdSecret = s
-	b.details.ProviderConfigOldFormat = cmapi.ACMEIssuerDNS01Provider{
-		Name: "dummy-provider",
-		Cloudflare: &cmapi.ACMEIssuerDNS01ProviderCloudflare{
-			Email: b.cf.Email,
-			APIKey: cmapi.SecretKeySelector{
-				LocalObjectReference: cmapi.LocalObjectReference{
-					Name: b.createdSecret.Name,
-				},
-				Key: "api-key",
-			},
-		},
-	}
 	b.details.ProviderConfig = cmapi.ACMEChallengeSolverDNS01{
 		Cloudflare: &cmapi.ACMEIssuerDNS01ProviderCloudflare{
 			Email: b.cf.Email,
